requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
groq
python-dotenv
numpy
